requirements.txt
openai
youtube-transcript-api
langchain
tiktoken
streamlit
python-dotenv
pytube
